Move InlineEdit reference to a more fitting place
(cherry picked from commit 2c115f51c67afd9e51b66c97778097ebbf773b8d)
