+++ src/components/DoctorCard/ReportError/styles/index.js
@@ -1,0 +1,1 @@
+export * as InlineEdit from './InlineEdit';
